+++ src/regex/classes.js
@@ -1,8 +1,0 @@
-/**
- * Talisman regex/classes
- * =======================
- *
- * Compilation of useful Regex classes.
- */
-export const SIMPLE_QUOTES = '’‘`‛\'';
-export const DOUBLE_QUOTES = '«»„‟“”"';