+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fixing a bug with `clustering/k-means` where k would be superior to the number of vectors.
 * Fixing a bug with `clustering/k-means` `initialCentroids` options.
 * Fixing a bug with `clustering/k-means` where a vector could end up in several clusters.
+* Dropping the internal `regex/classes` namespace.
 
 ## 0.11.0